Moved ImageReadLibraryWrapperUnitTest file.
