+++ PhotoViewerUnitTest/TestData/Penguins.NEF
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:435d67f7177a5d821786cbba247952d9f87e6a6af177df9dc4cbb37052ba6004
+size 19839360
